+++ src/js/components/error.test.tsx
@@ -59,14 +59,11 @@ describe('ErrorDisplay', (): void => {
       runtime: {},
       storage: {
         local: {
-          set: (obj: { [key: string]: string }, cb: () => void): void => {
+          set: (obj: { [key: string]: string }): Promise<void> => {
             Object.assign(localData, obj);
-            cb();
+            return Promise.resolve();
           },
-          get: (
-            keys: string[],
-            cb: (items: { [key: string]: string }) => void,
-          ): void => {
+          get: (keys: string[]): Promise<{ [key: string]: string }> => {
             const res: { [key: string]: string } = {};
             for (const key of keys) {
               const value = localData[key];
@@ -74,11 +71,11 @@ describe('ErrorDisplay', (): void => {
                 res[key] = value;
               }
             }
-            cb(res);
+            return Promise.resolve(res);
           },
-          remove: (key: string, cb: () => void): void => {
+          remove: (key: string): Promise<void> => {
             delete localData[key];
-            cb();
+            return Promise.resolve();
           },
         },
         onChanged: {